inject room database
