--- app/src/main/java/com/example/gear_kotlin/di/ViewModelFactoryModule.kt
@@ -1,15 +1,0 @@
-package com.example.gear_kotlin.di
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.example.gear_kotlin.viewmodel.ThirdFragmentViewModel
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class ViewModelFactoryModule {
-
-    @Binds
-    abstract fun bindViewModelFactory(factory: DaggerViewModelFactory): ViewModelProvider.Factory
-}
