+++ app/src/main/java/com/example/gear_kotlin/di/ViewModelKey.kt
@@ -8,5 +8,6 @@ import kotlin.reflect.KClass
 @Target(AnnotationTarget.FUNCTION,
     AnnotationTarget.PROPERTY_GETTER,
     AnnotationTarget.PROPERTY_SETTER)
+@Retention(AnnotationRetention.RUNTIME)
 @MapKey
 annotation class ViewModelKey(val value: KClass<out ViewModel>)
